Fix fileinfo filepos not being byteswapped (my fault)

+++ fbdoom/w_wad.c
@@ -216,7 +216,7 @@ wad_file_t *W_AddFile(char *filename)
     for (i=startlump; i<numlumps; ++i)
     {
 		lump_p->wad_file = wad_file;
-		lump_p->position = filerover->filepos;
+		lump_p->position = LONG(filerover->filepos);
 		lump_p->size = LONG(filerover->size);
 		lump_p->cache = NULL;
 		strncpy(lump_p->name, filerover->name, 8);
